Refactored optimization methods into different files.
