+++ sindy/optimizers/__init__.py
@@ -1,1 +1,5 @@
-from .optimizers import *
+from .base import BaseOptimizer
+from .stlsq import STLSQ
+from .sr3 import SR3
+from .lasso import LASSO
+from .elastic_net import ElasticNet
